DIAM-172 Add Diamante Entities to ACL DIAM-173 Cover Code with ACL Verification

+++ Ticket/Api/CommentServiceImpl.php
@@ -24,6 +24,8 @@ use Eltrino\DiamanteDeskBundle\Ticket\Api\Internal\UserService;
 use Eltrino\DiamanteDeskBundle\Ticket\Model\CommentRepository;
 use Eltrino\DiamanteDeskBundle\Ticket\Model\TicketRepository;
 use Oro\Bundle\UserBundle\Entity\User;
+use Oro\Bundle\SecurityBundle\SecurityFacade;
+use Oro\Bundle\SecurityBundle\Exception\ForbiddenException;
 
 class CommentServiceImpl implements CommentService
 {
@@ -52,18 +54,25 @@ class CommentServiceImpl implements CommentService
      */
     private $attachmentService;
 
+    /**
+     * @var \Oro\Bundle\SecurityBundle\SecurityFacade
+     */
+    private $securityFacade;
+
     public function __construct(
         TicketRepository $ticketRepository,
         CommentRepository $commentRepository,
         CommentFactory $commentFactory,
         UserService $userService,
-        AttachmentService $attachmentService
+        AttachmentService $attachmentService,
+        SecurityFacade $securityFacade
     ) {
         $this->ticketRepository = $ticketRepository;
         $this->commentRepository = $commentRepository;
         $this->commentFactory = $commentFactory;
         $this->userService = $userService;
         $this->attachmentService = $attachmentService;
+        $this->securityFacade = $securityFacade;
     }
 
     /**
@@ -76,6 +85,10 @@ class CommentServiceImpl implements CommentService
      */
     public function postNewCommentForTicket($content, $ticketId, $authorId, array $attachmentsInput = null)
     {
+        if (!$this->securityFacade->isGranted('CREATE', 'Entity:EltrinoDiamanteDeskBundle:Comment')) {
+            throw new ForbiddenException("Not enough permissions.");
+        }
+
         $ticket = $this->ticketRepository->get($ticketId);
 
         if (is_null($ticket)) {
@@ -123,9 +136,15 @@ class CommentServiceImpl implements CommentService
     public function updateTicketComment($commentId, $content, array $attachmentsInput = null)
     {
         $comment = $this->commentRepository->get($commentId);
+
         if (is_null($comment)) {
             throw new \RuntimeException('Comment loading failed, comment not found.');
         }
+
+        if (!$this->securityFacade->isGranted('EDIT', $comment)) {
+            throw new ForbiddenException("Not enough permissions.");
+        }
+
         $comment->updateContent($content);
         if ($attachmentsInput) {
             $this->attachmentService->createAttachmentsForItHolder($attachmentsInput, $comment);
@@ -140,9 +159,15 @@ class CommentServiceImpl implements CommentService
     public function deleteTicketComment($commentId)
     {
         $comment = $this->commentRepository->get($commentId);
+
         if (is_null($comment)) {
             throw new \RuntimeException('Comment loading failed, comment not found.');
         }
+
+        if (!$this->securityFacade->isGranted('DELETE', $comment)) {
+            throw new ForbiddenException("Not enough permissions.");
+        }
+
         $this->commentRepository->remove($comment);
     }
 
@@ -156,9 +181,15 @@ class CommentServiceImpl implements CommentService
     public function removeAttachmentFromComment($commentId, $attachmentId)
     {
         $comment = $this->commentRepository->get($commentId);
-        if (!$comment) {
+
+        if (is_null($comment)) {
             throw new \RuntimeException('Comment loading failed, comment not found.');
         }
+
+        if (!$this->securityFacade->isGranted('EDIT', $comment)) {
+            throw new ForbiddenException("Not enough permissions.");
+        }
+
         $attachment = $comment->getAttachment($attachmentId);
         if (!$attachment) {
             throw new \RuntimeException('Attachment loading failed. Comment has no such attachment.');
@@ -171,14 +202,16 @@ class CommentServiceImpl implements CommentService
     public static function create(
         \Doctrine\ORM\EntityManager $em,
         UserService $userService,
-        AttachmentService $attachmentService
+        AttachmentService $attachmentService,
+        SecurityFacade $securityFacade
     ) {
         return new CommentServiceImpl(
             $em->getRepository('Eltrino\DiamanteDeskBundle\Entity\Ticket'),
             $em->getRepository('Eltrino\DiamanteDeskBundle\Entity\Comment'),
             new CommentFactory(),
             $userService,
-            $attachmentService
+            $attachmentService,
+            $securityFacade
         );
     }
 }
